+++ integrations/sdk/exceptions.py
@@ -1,0 +1,17 @@
+"""Exceptions raised by Integration SDK providers."""
+
+
+class IntegrationError(Exception):
+    pass
+
+
+class NotSupported(IntegrationError):
+    pass
+
+
+class AuthFailed(IntegrationError):
+    pass
+
+
+class RateLimited(IntegrationError):
+    pass